JAMES-2997 Renable and relocate search tests
